Move annotations
