+++ core/contract/src/main/java/org/eclipse/dataspaceconnector/contract/ContractNegotiationCommandExtension.java
@@ -14,7 +14,7 @@
 package org.eclipse.dataspaceconnector.contract;
 
 import org.eclipse.dataspaceconnector.contract.negotiation.command.handlers.CancelNegotiationCommandHandler;
-import org.eclipse.dataspaceconnector.core.CoreExtension;
+import org.eclipse.dataspaceconnector.spi.system.CoreExtension;
 import org.eclipse.dataspaceconnector.core.base.CommandHandlerRegistryImpl;
 import org.eclipse.dataspaceconnector.spi.command.CommandHandlerRegistry;
 import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNegotiationStore;
@@ -30,10 +30,10 @@ import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
 @CoreExtension
 @Provides({CommandHandlerRegistry.class})
 public class ContractNegotiationCommandExtension implements ServiceExtension {
-    
+
     @Inject
     private ContractNegotiationStore store;
-    
+
     @Override
     public void initialize(ServiceExtensionContext context) {
         CommandHandlerRegistry registry = context.getService(CommandHandlerRegistry.class, true);
@@ -41,12 +41,12 @@ public class ContractNegotiationCommandExtension implements ServiceExtension {
             registry = new CommandHandlerRegistryImpl();
             context.registerService(CommandHandlerRegistry.class, registry);
         }
-        
+
         registerDefaultCommandHandlers(registry);
     }
-    
+
     private void registerDefaultCommandHandlers(CommandHandlerRegistry registry) {
         registry.register(new CancelNegotiationCommandHandler(store));
     }
-    
+
 }
